src/requirements.txt
backports.ssl-match-hostname==3.4.0.2
blivet==0.61.15.65
Brlapi==0.6.0
cffi==1.6.0
chardet==2.2.1
configobj==4.7.2
configshell-fb==1.1.23
coverage==3.6b3
cryptography==1.7.2
cupshelpers==1.0
decorator==3.4.0
di==0.3
dj-database-url==0.4.2
django-crispy-forms==1.7.0
dnspython==1.12.0
enum34==1.0.4
ethtool==0.8
firstboot==19.5
fros==1.0
gssapi==1.2.0
gunicorn==19.7.1
html2text==2016.9.19
idna==2.4
iniparse==0.4
initial-setup==0.3.9.40
ipaclient==4.5.0
ipaddress==1.0.16
ipalib==4.5.0
ipaplatform==4.5.0
ipapython==4.5.0
IPy==0.75
javapackages==1.0.0
jwcrypto==0.2.1
kitchen==1.1.1
kmod==0.1
langtable==0.0.31
lxml==3.2.1
netaddr==0.7.5
netifaces==0.10.4
ntplib==0.3.2
perf==0.1
Pillow==5.0.0
ply==3.4
policycoreutils-default-encoding==0.1
psycopg2==2.7.3.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pycups==1.9.63
pycurl==7.19.0
pygobject==3.22.0
pygpgme==0.3
pyinotify==0.9.4
pykickstart==1.99.66.12
pyliblzma==0.5.3
pyOpenSSL==0.13.1
pyparsing==1.5.6
pyparted==3.9
pysmbc==1.0.13
python-augeas==0.5.0
python-dateutil==1.5
python-dmidecode==3.12.2
python-ldap==2.4.15
python-linux-procfs==0.4.9
python-meh==0.25.2
python-nmap==0.6.1
python-nss==0.16.0
python-yubico==1.2.3
pytz==2016.10
pyudev==0.15
pyusb==1.0.0b1
pyxattr==0.5.1
qrcode==5.0.1
requests==2.6.0
rtslib-fb==2.1.63
schedutils==0.4
seobject==0.1
sepolicy==1.1
setroubleshoot==1.1
six==1.9.0
slip==0.4.0
slip.dbus==0.4.0
SSSDConfig==1.15.2
targetcli-fb===2.1.fb46
urlgrabber==3.10
urllib3==1.10.2
urwid==1.1.1
virtualenv==15.1.0
yum-langpacks==0.4.2
yum-metadata-parser==1.1.4
